% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_package_installed.R
\name{check_if_package_installed}
\alias{check_if_package_installed}
\title{Check If a Package is Installed}
\usage{
check_if_package_installed(package, type)
}
\arguments{
\item{package}{The name of the package to check.}

\item{type}{The type for which the package needs to be available.}
}
\value{
Invisible `TRUE` if the package is installed, otherwise stops
        execution with an error message advising the installation of the package.
        Since the function is designed to stop if the package is not found,
        it does not explicitly return a value upon successful completion.
}
\description{
Checks if a specified package is installed and available for use. If the package
is not installed, the function stops execution and prompts the user to install
the package.
}
