% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_supported_type.R
\name{check_supported_type}
\alias{check_supported_type}
\title{Check if a Dataset Type is Supported}
\usage{
check_supported_type(type)
}
\arguments{
\item{type}{A character string specifying the dataset type to check.}
}
\value{
Does not return a value; instead, it either passes silently if the
type is supported or stops execution with an error message if the type is
unsupported.
}
\description{
Checks if a given dataset type is supported by verifying against a list of
all supported dataset types from different domains. If the specified type is
not supported, execution stops and an error message listing all supported
types is returned.
}
