% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wrds_dummy_database.R
\name{create_wrds_dummy_database}
\alias{create_wrds_dummy_database}
\title{Create WRDS Dummy Database}
\usage{
create_wrds_dummy_database(
  path,
  url = paste0("https://github.com/tidy-finance/website/tree/main/blog/",
    "tidy-finance-dummy-data/data/tidy_finance.sqlite")
)
}
\arguments{
\item{path}{The file path where the SQLite database should be saved.}

\item{url}{The URL where the SQLite database is stored.}
}
\value{
Invisible \code{NULL}. Side effect: downloads a file to the specified path.
}
\description{
Downloads the WRDS dummy database from the respective Tidy Finance GitHub
repository and saves it to the specified path. If the file already exists,
the user is prompted before it is replaced.
}
\examples{
path <- paste0(tempdir(), "/tidy_finance_r.sqlite")
create_wrds_dummy_database(path)
}
\seealso{
Other WRDS functions: 
\code{\link{disconnect_connection}()},
\code{\link{download_data_wrds}()},
\code{\link{download_data_wrds_ccm_links}()},
\code{\link{download_data_wrds_compustat}()},
\code{\link{download_data_wrds_crsp}()},
\code{\link{download_data_wrds_fisd}()},
\code{\link{download_data_wrds_trace_enhanced}()},
\code{\link{get_wrds_connection}()},
\code{\link{set_wrds_credentials}()}
}
\concept{WRDS functions}
