% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect_connection.R
\name{disconnect_connection}
\alias{disconnect_connection}
\title{Disconnect Database Connection}
\usage{
disconnect_connection(con)
}
\arguments{
\item{con}{A database connection object created by DBI::dbConnect or any
similar function that establishes a connection to a database.}
}
\value{
\code{TRUE}, invisibly. Throws an error if the disconnection fails.
}
\description{
Safely disconnects an established database connection using the DBI package.
}
\examples{
\dontrun{
con <- get_wrds_connection()
disconnect_connection(con)
}
}
\seealso{
Other WRDS functions:
\code{\link[=download_data_wrds]{download_data_wrds()}},
\code{\link[=download_data_wrds_ccm_links]{download_data_wrds_ccm_links()}},
\code{\link[=download_data_wrds_compustat]{download_data_wrds_compustat()}},
\code{\link[=download_data_wrds_crsp]{download_data_wrds_crsp()}},
\code{\link[=download_data_wrds_fisd]{download_data_wrds_fisd()}},
\code{\link[=download_data_wrds_trace_enhanced]{download_data_wrds_trace_enhanced()}},
\code{\link[=get_wrds_connection]{get_wrds_connection()}},
\code{\link[=set_wrds_credentials]{set_wrds_credentials()}}
}
\concept{WRDS functions}
