% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wrds_connection.R
\name{disconnection_connection}
\alias{disconnection_connection}
\title{Disconnect Database Connection}
\usage{
disconnection_connection(con)
}
\arguments{
\item{con}{A database connection object created by DBI::dbConnect or any similar function that
establishes a connection to a database.}
}
\value{
A logical value: `TRUE` if disconnection was successful, `FALSE` otherwise.
}
\description{
This function safely disconnects an established database connection using the DBI package.
}
