% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random_user_agent.R
\name{get_random_user_agent}
\alias{get_random_user_agent}
\title{Get a Random User Agent}
\usage{
get_random_user_agent()
}
\value{
A character string representing a randomly selected user agent.
}
\description{
This function selects and returns a random user agent string from a predefined list.
The list contains user agents for various operating systems and browsers, including
Windows, macOS, Linux, Android, iPhone, Chrome, Safari, Firefox, and Edge.
}
