% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_supported_datasets.R
\name{list_supported_datasets_ff_legacy}
\alias{list_supported_datasets_ff_legacy}
\title{List Supported Legacy Fama-French Datasets}
\usage{
list_supported_datasets_ff_legacy()
}
\value{
A tibble with columns: \code{type} (the type of dataset), \code{dataset_name}
(a descriptive name of the dataset), \code{file_url} (the path of the source
ZIP relative to Kenneth French's data library), and \code{domain} (the domain
to which the dataset belongs, always "Fama-French").
}
\description{
Returns a tibble with the legacy names of initially supported
Fama-French datasets, including their names and frequencies (daily,
weekly, monthly). Each dataset type is associated with a specific Fama-French
model (e.g., 3 factors, 5 factors). Additionally, it annotates each dataset
with the domain "Fama-French". Not included in the exported
\code{list_supported_datasets()} function.
}
\seealso{
Other utility functions:
\code{\link[=create_summary_statistics]{create_summary_statistics()}},
\code{\link[=get_available_huggingface_files]{get_available_huggingface_files()}},
\code{\link[=list_supported_datasets]{list_supported_datasets()}},
\code{\link[=list_supported_datasets_ff]{list_supported_datasets_ff()}},
\code{\link[=list_supported_datasets_macro_predictors]{list_supported_datasets_macro_predictors()}},
\code{\link[=list_supported_datasets_other]{list_supported_datasets_other()}},
\code{\link[=list_supported_datasets_pseudo]{list_supported_datasets_pseudo()}},
\code{\link[=list_supported_datasets_wrds]{list_supported_datasets_wrds()}},
\code{\link[=list_supported_indexes]{list_supported_indexes()}},
\code{\link[=list_supported_jkp_factors]{list_supported_jkp_factors()}},
\code{\link[=list_tidy_finance_chapters]{list_tidy_finance_chapters()}},
\code{\link[=open_tidy_finance_website]{open_tidy_finance_website()}},
\code{\link[=trim]{trim()}},
\code{\link[=validate_dates]{validate_dates()}},
\code{\link[=winsorize]{winsorize()}}
}
\concept{utility functions}
