% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_supported_datasets.R
\name{list_supported_datasets_macro_predictors}
\alias{list_supported_datasets_macro_predictors}
\title{List Supported Macro Predictor Datasets}
\usage{
list_supported_datasets_macro_predictors()
}
\value{
A tibble with columns: \code{type} (the type of dataset),
\code{dataset_name} (the file name of the dataset, which is the same for
all types), and \code{domain} (the domain to which the dataset belongs,
always "Goyal-Welch").
}
\description{
Returns a tibble with the supported macro predictor datasets
provided by Goyal-Welch, including their frequencies (monthly, quarterly,
annual). All datasets reference the same source file,
"PredictorData2022.xlsx" for the year 2022. Additionally, it annotates
each dataset with the domain "Goyal-Welch".
}
\seealso{
Other utility functions:
\code{\link[=create_summary_statistics]{create_summary_statistics()}},
\code{\link[=get_available_huggingface_files]{get_available_huggingface_files()}},
\code{\link[=list_supported_datasets]{list_supported_datasets()}},
\code{\link[=list_supported_datasets_ff]{list_supported_datasets_ff()}},
\code{\link[=list_supported_datasets_ff_legacy]{list_supported_datasets_ff_legacy()}},
\code{\link[=list_supported_datasets_other]{list_supported_datasets_other()}},
\code{\link[=list_supported_datasets_pseudo]{list_supported_datasets_pseudo()}},
\code{\link[=list_supported_datasets_wrds]{list_supported_datasets_wrds()}},
\code{\link[=list_supported_indexes]{list_supported_indexes()}},
\code{\link[=list_supported_jkp_factors]{list_supported_jkp_factors()}},
\code{\link[=list_tidy_finance_chapters]{list_tidy_finance_chapters()}},
\code{\link[=open_tidy_finance_website]{open_tidy_finance_website()}},
\code{\link[=trim]{trim()}},
\code{\link[=validate_dates]{validate_dates()}},
\code{\link[=winsorize]{winsorize()}}
}
\concept{utility functions}
