% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_supported_datasets.R
\name{list_supported_datasets_wrds}
\alias{list_supported_datasets_wrds}
\title{List Supported WRDS Datasets}
\usage{
list_supported_datasets_wrds()
}
\value{
A tibble with columns: \code{type} (the type of dataset),
\code{dataset_name} (the file name of the dataset), and \code{domain} (the
domain to which the dataset belongs, always "WRDS").
}
\description{
Returns a tibble with the supported datasets provided via WRDS.
Additionally, it annotates each dataset with the domain "WRDS".
}
\seealso{
Other utility functions:
\code{\link[=create_summary_statistics]{create_summary_statistics()}},
\code{\link[=get_available_huggingface_files]{get_available_huggingface_files()}},
\code{\link[=list_supported_datasets]{list_supported_datasets()}},
\code{\link[=list_supported_datasets_ff]{list_supported_datasets_ff()}},
\code{\link[=list_supported_datasets_ff_legacy]{list_supported_datasets_ff_legacy()}},
\code{\link[=list_supported_datasets_macro_predictors]{list_supported_datasets_macro_predictors()}},
\code{\link[=list_supported_datasets_other]{list_supported_datasets_other()}},
\code{\link[=list_supported_datasets_pseudo]{list_supported_datasets_pseudo()}},
\code{\link[=list_supported_indexes]{list_supported_indexes()}},
\code{\link[=list_supported_jkp_factors]{list_supported_jkp_factors()}},
\code{\link[=list_tidy_finance_chapters]{list_tidy_finance_chapters()}},
\code{\link[=open_tidy_finance_website]{open_tidy_finance_website()}},
\code{\link[=trim]{trim()}},
\code{\link[=validate_dates]{validate_dates()}},
\code{\link[=winsorize]{winsorize()}}
}
\concept{utility functions}
