% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_supported_types.R
\name{list_supported_types_stocks}
\alias{list_supported_types_stocks}
\title{List Supported Stock Data Types}
\usage{
list_supported_types_stocks()
}
\value{
A tibble with columns \code{type} and \code{dataset_name}, where \code{type} indicates the code used to specify the data source and \code{dataset_name} provides the name of the data source.
}
\description{
Returns a tibble listing the supported stock data types and their corresponding dataset names.
}
