% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tidy_finance_chapters.R
\name{list_tidy_finance_chapters}
\alias{list_tidy_finance_chapters}
\title{List Chapters of Tidy Finance}
\usage{
list_tidy_finance_chapters()
}
\value{
A character vector where each element is the name of a chapter
available in the Tidy Finance resource. These names correspond to specific
chapters in Tidy Finance with R.
}
\description{
Returns a character vector containing the names of the chapters available in
the Tidy Finance resource. It provides a quick reference to the
various topics covered.
}
\examples{
list_tidy_finance_chapters()

}
\seealso{
Other utility functions:
\code{\link[=create_summary_statistics]{create_summary_statistics()}},
\code{\link[=get_available_huggingface_files]{get_available_huggingface_files()}},
\code{\link[=list_supported_datasets]{list_supported_datasets()}},
\code{\link[=list_supported_datasets_ff]{list_supported_datasets_ff()}},
\code{\link[=list_supported_datasets_ff_legacy]{list_supported_datasets_ff_legacy()}},
\code{\link[=list_supported_datasets_macro_predictors]{list_supported_datasets_macro_predictors()}},
\code{\link[=list_supported_datasets_other]{list_supported_datasets_other()}},
\code{\link[=list_supported_datasets_pseudo]{list_supported_datasets_pseudo()}},
\code{\link[=list_supported_datasets_wrds]{list_supported_datasets_wrds()}},
\code{\link[=list_supported_indexes]{list_supported_indexes()}},
\code{\link[=list_supported_jkp_factors]{list_supported_jkp_factors()}},
\code{\link[=open_tidy_finance_website]{open_tidy_finance_website()}},
\code{\link[=trim]{trim()}},
\code{\link[=validate_dates]{validate_dates()}},
\code{\link[=winsorize]{winsorize()}}
}
\concept{utility functions}
