% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_tidy_finance_website.R
\name{open_tidy_finance_website}
\alias{open_tidy_finance_website}
\title{Open Tidy Finance Website or Specific Chapter in Browser}
\usage{
open_tidy_finance_website(chapter = NULL)
}
\arguments{
\item{chapter}{An optional character string specifying the chapter to open.
If \code{NULL} (the default), the function opens the main page of Tidy Finance
with R. If a chapter name is provided (e.g., "beta-estimation"), the
function opens the corresponding chapter's page (e.g.,
"beta-estimation.html"). If the chapter name does not exist, then the
function opens the main page.}
}
\value{
Invisible \code{NULL}. The function is called for its side effect of
opening a web page.
}
\description{
Opens the main Tidy Finance website or a specific chapter within the site in
the user's default web browser. If a chapter is specified, the function
constructs the URL to access the chapter directly.
}
\examples{
\dontrun{
open_tidy_finance_website()
open_tidy_finance_website("beta-estimation")
}

}
\seealso{
Other utility functions:
\code{\link[=create_summary_statistics]{create_summary_statistics()}},
\code{\link[=get_available_huggingface_files]{get_available_huggingface_files()}},
\code{\link[=list_supported_datasets]{list_supported_datasets()}},
\code{\link[=list_supported_datasets_ff]{list_supported_datasets_ff()}},
\code{\link[=list_supported_datasets_ff_legacy]{list_supported_datasets_ff_legacy()}},
\code{\link[=list_supported_datasets_macro_predictors]{list_supported_datasets_macro_predictors()}},
\code{\link[=list_supported_datasets_other]{list_supported_datasets_other()}},
\code{\link[=list_supported_datasets_pseudo]{list_supported_datasets_pseudo()}},
\code{\link[=list_supported_datasets_wrds]{list_supported_datasets_wrds()}},
\code{\link[=list_supported_indexes]{list_supported_indexes()}},
\code{\link[=list_supported_jkp_factors]{list_supported_jkp_factors()}},
\code{\link[=list_tidy_finance_chapters]{list_tidy_finance_chapters()}},
\code{\link[=trim]{trim()}},
\code{\link[=validate_dates]{validate_dates()}},
\code{\link[=winsorize]{winsorize()}}
}
\concept{utility functions}
