% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_wrds_credentials.R
\name{set_wrds_credentials}
\alias{set_wrds_credentials}
\title{Set WRDS Credentials}
\usage{
set_wrds_credentials()
}
\value{
Invisibly returns \code{TRUE}. Displays messages to the user based on
their input and actions taken.
}
\description{
Prompts the user to input their WRDS (Wharton Research Data Services)
username and password, and stores these credentials in a \code{.Renviron} file.
The user can choose to store the \code{.Renviron} file in either the project
directory or the home directory. If the \code{.Renviron} file already contains
WRDS credentials, the user will be asked if they want to overwrite the
existing credentials. Additionally, the user has the option to add the
\code{.Renviron} file to the .gitignore file to prevent it from being tracked by
version control.
}
\examples{
\dontrun{
set_wrds_credentials()
}
}
\seealso{
Other WRDS functions:
\code{\link[=disconnect_connection]{disconnect_connection()}},
\code{\link[=download_data_wrds]{download_data_wrds()}},
\code{\link[=download_data_wrds_ccm_links]{download_data_wrds_ccm_links()}},
\code{\link[=download_data_wrds_compustat]{download_data_wrds_compustat()}},
\code{\link[=download_data_wrds_crsp]{download_data_wrds_crsp()}},
\code{\link[=download_data_wrds_fisd]{download_data_wrds_fisd()}},
\code{\link[=download_data_wrds_trace_enhanced]{download_data_wrds_trace_enhanced()}},
\code{\link[=get_wrds_connection]{get_wrds_connection()}}
}
\concept{WRDS functions}
