% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_long_short_returns.R
\name{compute_long_short_returns}
\alias{compute_long_short_returns}
\title{Compute long-short portfolio returns}
\usage{
compute_long_short_returns(portfolio_returns, direction = "top_minus_bottom")
}
\arguments{
\item{portfolio_returns}{Some dataset...}
}
\value{
A tibble with long-short returns.
}
\description{
Compute long-short portfolio returns
}
\examples{
\donttest{
  compute_long_short_returns(portfolio_returns)
}

}
