% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wrds_dummy_database.R
\name{create_wrds_dummy_database}
\alias{create_wrds_dummy_database}
\title{Create WRDS Dummy Database}
\usage{
create_wrds_dummy_database(path)
}
\arguments{
\item{path}{The file path where the SQLite database should be saved. If not
provided, the default path is "data/tidy_finance_r.sqlite".}
}
\value{
Invisible NULL. Side effect: downloads a file to the specified path.
}
\description{
Downloads the WRDS dummy database from the respective Tidy Finance GitHub
repository and saves it to the specified path. If the file already exists,
the user is prompted before it is replaced.
}
\examples{
path <- paste0(tempdir(), "/tidy_finance_r.sqlite")
create_wrds_dummy_database(path)

}
