% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_options.R
\name{data_options}
\alias{data_options}
\title{Create Data Options}
\usage{
data_options(
  id = "permno",
  date = "date",
  exchange = "exchange",
  mktcap_lag = "mktcap_lag",
  ret_excess = "ret_excess",
  portfolio = "portfolio",
  siccd = "siccd",
  price = "prc_adj",
  listing_age = "listing_age",
  be = "be",
  ...
)
}
\arguments{
\item{id}{A character string representing the entity variable (defaults to
\code{"permno"}).}

\item{date}{A character string representing the date variable (defaults to
\code{"date"}).}

\item{exchange}{A character string representing the exchange variable
(defaults to \code{"exchange"}).}

\item{mktcap_lag}{A character string representing the market capitalization
lag variable (defaults to \code{"mktcap_lag"}).}

\item{ret_excess}{A character string representing the excess return variable
(defaults to \code{"ret_excess"}).}

\item{portfolio}{A character string representing the portfolio variable
(defaults to \code{"portfolio"}).}

\item{siccd}{A character string representing the Standard Industrial
Classification code variable (defaults to \code{"siccd"}).}

\item{price}{A character string representing the (adjusted) price variable
(defaults to \code{"prc_adj"}).}

\item{listing_age}{A character string representing the listing age variable
(defaults to \code{"listing_age"}).}

\item{be}{A character string representing the book equity variable (defaults
to \code{"be"}).}

\item{...}{Additional arguments to be included in the data options list.}
}
\value{
A list of class \code{tidyfinance_data_options} containing the specified
data options.
}
\description{
Creates a list of data options of class \code{tidyfinance_data_options} used for
TidyFinance-related functions. These options map the specific data variables
to the TidyFinance naming conventions, allowing functions to flexibly work
with different datasets by specifying the relevant column names. Additional
options can be passed through \code{...}.
}
\examples{
data_options(
  id = "permno",
  date = "date",
  exchange = "exchange"
)

}
\seealso{
Other portfolio functions: 
\code{\link{assign_portfolio}()},
\code{\link{breakpoint_options}()},
\code{\link{compute_breakpoints}()},
\code{\link{compute_long_short_returns}()},
\code{\link{compute_portfolio_returns}()},
\code{\link{implement_portfolio_sort}()},
\code{\link{filter_options}()},
\code{\link{filter_sorting_data}()},
\code{\link{portfolio_sort_options}()}
}
\concept{portfolio functions}
