% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_options.R
\name{data_options}
\alias{data_options}
\title{Create Data Options}
\usage{
data_options(
  id = "permno",
  date = "date",
  exchange = "exchange",
  mktcap_lag = "mktcap_lag",
  ret_excess = "ret_excess",
  portfolio = "portfolio",
  ...
)
}
\arguments{
\item{id}{A character string representing the entity variable (defaults to
\code{"permno"}).}

\item{date}{A character string representing the date variable (defaults to
\code{"date"}).}

\item{exchange}{A character string representing the exchange variable
(defaults to \code{"exchange"}).}

\item{mktcap_lag}{A character string representing the market capitalization
lag variable (defaults to \code{"mktcap_lag"}).}

\item{ret_excess}{A character string representing the excess return variable
(defaults to \code{"ret_excess"}).}

\item{portfolio}{A character string representing the portfolio variable
(defaults to \code{"portfolio"}).}

\item{...}{Additional arguments to be included in the data options list.}
}
\value{
A list of class \code{tidyfinance_data_options} containing the specified
data options.
}
\description{
Creates a list of data options of class \code{tidyfinance_data_options} used for
TidyFinance-related functions. These options map the specific data variables
to the TidyFinance naming conventions, allowing functions to flexibly work
with different datasets by specifying the relevant column names. The
function accepts key parameters such as \code{id}, \code{date}, \code{exchange},
\code{mktcap_lag}, \code{ret_excess}, \code{portfolio}, along with other additional
options passed through \code{...}.
}
\examples{
data_options(
  id = "permno",
  date = "date",
  exchange = "exchange"
)

}
\seealso{
Other portfolio functions: 
\code{\link{assign_portfolio}()},
\code{\link{breakpoint_options}()},
\code{\link{compute_breakpoints}()},
\code{\link{compute_long_short_returns}()},
\code{\link{compute_portfolio_returns}()}
}
\concept{portfolio functions}
