% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect_connection.R
\name{disconnect_connection}
\alias{disconnect_connection}
\title{Disconnect Database Connection}
\usage{
disconnect_connection(con)
}
\arguments{
\item{con}{A database connection object created by DBI::dbConnect or any
similar function that establishes a connection to a database.}
}
\value{
A logical value: \code{TRUE} if disconnection was successful, \code{FALSE}
otherwise.
}
\description{
Safely disconnects an established database connection using the DBI package.
}
