% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_macro_predictors.R
\name{download_data_macro_predictors}
\alias{download_data_macro_predictors}
\title{Download and Process Macro Predictor Data}
\usage{
download_data_macro_predictors(
  dataset = NULL,
  start_date = NULL,
  end_date = NULL,
  type = deprecated(),
  sheet_id = "1bM7vCWd3WOt95Sf9qjLPZjoiafgF_8EG"
)
}
\arguments{
\item{dataset}{The dataset to download ("monthly", "quarterly", "annual").}

\item{start_date}{Optional. A character string or Date object in "YYYY-MM-DD" format
specifying the start date for the data. If not provided, the full dataset is returned.}

\item{end_date}{Optional. A character string or Date object in "YYYY-MM-DD" format
specifying the end date for the data. If not provided, the full dataset is returned.}

\item{type}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{dataset} instead.}

\item{sheet_id}{The Google Sheets ID from which to download the dataset, with the
default "1bM7vCWd3WOt95Sf9qjLPZjoiafgF_8EG".}
}
\value{
A tibble with processed data, filtered by the specified date range
and including financial metrics.
}
\description{
Downloads and processes macroeconomic predictor data based on the specified
dataset (monthly, quarterly, or annual), date range, and source URL. The
function downloads the data from a Google Sheets export link. It
processes the raw data into a structured format, calculating additional
financial metrics and filtering by the specified date range.
}
\examples{
\donttest{
  download_data_macro_predictors("monthly")
  download_data_macro_predictors("quarterly", "2000-01-01", "2020-12-31")
}

}
