% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_betas.R
\name{estimate_betas}
\alias{estimate_betas}
\title{Estimate Rolling Betas}
\usage{
estimate_betas(
  data,
  model,
  months_lookback,
  min_obs = round(months_lookback * 0.8, 0),
  n_cores = 1
)
}
\arguments{
\item{data}{A tibble containing the data with columns \code{date}, \code{permno}, and other variables used in the model.}

\item{model}{A formula representing the model to be estimated (e.g., \code{ret_excess ~ mkt_excess + smb + hml}).}

\item{months_lookback}{An integer specifying the number of months to look back when estimating the rolling model.}

\item{min_obs}{An integer specifying the minimum number of observations required to estimate the model.
Defaults to 80\% of \code{months_lookback}.}

\item{n_cores}{An integer specifying the number of cores to use for parallel processing. Defaults to 1.}
}
\value{
A tibble with the estimated betas for each time period.
}
\description{
This function estimates rolling betas for a given model using the provided data.
It supports parallel processing for faster computation using the \code{furrr} package.
}
\examples{
set.seed(1234)
data <- tibble::tibble(
  date = rep(seq.Date(from = as.Date("2020-01-01"), to = as.Date("2020-12-01"), by = "month"), each = 50),
  permno = rep(1:50, times = 12),
  ret_excess = rnorm(600, 0, 0.1),
  mkt_excess = rnorm(600, 0, 0.1),
  smb = rnorm(600, 0, 0.1),
  hml = rnorm(600, 0, 0.1),
)

estimate_betas(data,  "ret_excess ~ mkt_excess", 3)
estimate_betas(data,  "ret_excess ~ mkt_excess + smb + hml", 6)

}
