% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_supported_datasets.R
\name{list_supported_datasets_pseudo}
\alias{list_supported_datasets_pseudo}
\title{List Supported Pseudo WRDS Datasets}
\usage{
list_supported_datasets_pseudo()
}
\value{
A tibble with columns: \code{type}, \code{dataset_name}, and \code{domain}.
}
\description{
Returns a tibble with the supported pseudo WRDS datasets generated by
\code{download_data(domain = "pseudo", ...)}. Annotated with the domain
"pseudo".
}
\seealso{
Other utility functions: 
\code{\link{create_summary_statistics}()},
\code{\link{get_available_huggingface_files}()},
\code{\link{list_supported_datasets}()},
\code{\link{list_supported_datasets_ff}()},
\code{\link{list_supported_datasets_ff_legacy}()},
\code{\link{list_supported_datasets_macro_predictors}()},
\code{\link{list_supported_datasets_other}()},
\code{\link{list_supported_datasets_wrds}()},
\code{\link{list_supported_indexes}()},
\code{\link{list_tidy_finance_chapters}()},
\code{\link{open_tidy_finance_website}()},
\code{\link{trim}()},
\code{\link{validate_dates}()},
\code{\link{winsorize}()}
}
\concept{utility functions}
