% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_supported_datasets.R
\name{list_supported_datasets_wrds}
\alias{list_supported_datasets_wrds}
\title{List Supported WRDS Dataset Types}
\usage{
list_supported_datasets_wrds()
}
\value{
A tibble with columns: \code{type} (the type of dataset),
\code{dataset_name} (the file name of the dataset), and \code{domain} (the
domain to which the dataset belongs, always "WRDS").
}
\description{
Returns a tibble with the supported dataset types provided via WRDS.
Additionally, it annotates each dataset with the domain "WRDS".
}
\seealso{
Other utility functions: 
\code{\link{create_summary_statistics}()},
\code{\link{get_available_huggingface_files}()},
\code{\link{list_supported_datasets}()},
\code{\link{list_supported_datasets_ff}()},
\code{\link{list_supported_datasets_ff_legacy}()},
\code{\link{list_supported_datasets_macro_predictors}()},
\code{\link{list_supported_datasets_other}()},
\code{\link{list_supported_indexes}()},
\code{\link{list_tidy_finance_chapters}()},
\code{\link{open_tidy_finance_website}()},
\code{\link{trim}()},
\code{\link{validate_dates}()},
\code{\link{winsorize}()}
}
\concept{utility functions}
