% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_supported_indexes.R
\name{list_supported_indexes}
\alias{list_supported_indexes}
\title{List Supported Indexes}
\usage{
list_supported_indexes()
}
\value{
A tibble with three columns:
\describe{
\item{index}{The name of the financial index (e.g., "DAX", "S&P 500").}
\item{url}{The URL to the CSV file containing the holdings data for the index.}
\item{skip}{The number of lines to skip when reading the CSV file.}
}
}
\description{
This function returns a tibble containing information about supported financial indexes.
Each index is associated with a URL that points to a CSV file containing the holdings of the index.
Additionally, each index has a corresponding \code{skip} value, which indicates the number of lines
to skip when reading the CSV file.
}
\examples{
supported_indexes <- list_supported_indexes()
print(supported_indexes)

}
