% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{list_supported_types_macro_predictors}
\alias{list_supported_types_macro_predictors}
\title{List Supported Macro Predictor Dataset Types}
\usage{
list_supported_types_macro_predictors()
}
\value{
A tibble with columns: \code{type} (the type of dataset), \code{dataset_name}
(the file name of the dataset, which is the same for all types), and
\code{domain} (the domain to which the dataset belongs, always "Goyal-Welch").
}
\description{
This function returns a tibble with the supported macro predictor dataset
types provided by Goyal-Welch, including their frequencies (monthly,
quarterly, annual). All dataset types reference the same source file
"PredictorData2022.xlsx" for the year 2022. Additionally, it annotates each
dataset with the domain "Goyal-Welch".
}
