% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_supported_types.R
\name{list_supported_types_wrds}
\alias{list_supported_types_wrds}
\title{List Supported WRDS Dataset Types}
\usage{
list_supported_types_wrds()
}
\value{
A tibble with columns: \code{type} (the type of dataset), \code{dataset_name}
(the file name of the dataset), and \code{domain} (the domain to which the
dataset belongs, always "WRDS").
}
\description{
This function returns a tibble with the supported dataset types provided via
WRDS. Additionally, it annotates each dataset with the domain "WRDS".
}
