% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tidy_finance_chapters.R
\name{list_tidy_finance_chapters}
\alias{list_tidy_finance_chapters}
\title{List Chapters of Tidy Finance}
\usage{
list_tidy_finance_chapters()
}
\value{
A character vector where each element is the name of a chapter
available in the Tidy Finance resource. These names correspond to specific
chapters in Tidy Finance with R.
}
\description{
Returns a character vector containing the names of the chapters available in
the Tidy Finance resource. This function provides a quick reference to the
various topics covered.
}
\examples{
list_tidy_finance_chapters()
}
