% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_sort_options.R
\name{portfolio_sort_options}
\alias{portfolio_sort_options}
\title{Create Portfolio Sort Options}
\usage{
portfolio_sort_options(
  filter_options = NULL,
  breakpoint_options_main,
  breakpoint_options_secondary = NULL,
  ...
)
}
\arguments{
\item{filter_options}{A list of class \code{tidyfinance_filter_options} created
by \code{\link[=filter_options]{filter_options()}}, or \code{NULL} (the default, which applies no filters).}

\item{breakpoint_options_main}{A list of class
\code{tidyfinance_breakpoint_options} created by \code{\link[=breakpoint_options]{breakpoint_options()}},
specifying breakpoints for the primary sorting variable.}

\item{breakpoint_options_secondary}{A list of class
\code{tidyfinance_breakpoint_options} created by \code{\link[=breakpoint_options]{breakpoint_options()}},
specifying breakpoints for the secondary sorting variable, or \code{NULL}
(the default) for univariate sorts.}

\item{...}{Additional arguments to be included in the options list.}
}
\value{
A list of class \code{tidyfinance_portfolio_sort_options} containing
the specified options.
}
\description{
Creates a list of options of class \code{tidyfinance_portfolio_sort_options} that
bundles sample construction filters and breakpoint specifications for use
with \code{\link[=implement_portfolio_sort]{implement_portfolio_sort()}}.
}
\examples{
portfolio_sort_options(
  filter_options = filter_options(exclude_financials = TRUE),
  breakpoint_options_main = breakpoint_options(n_portfolios = 10)
)

}
\seealso{
Other portfolio functions: 
\code{\link{assign_portfolio}()},
\code{\link{breakpoint_options}()},
\code{\link{compute_breakpoints}()},
\code{\link{compute_long_short_returns}()},
\code{\link{compute_portfolio_returns}()},
\code{\link{data_options}()},
\code{\link{filter_options}()},
\code{\link{filter_sorting_data}()},
\code{\link{implement_portfolio_sort}()}
}
\concept{portfolio functions}
