% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyfinance-package.R
\docType{package}
\name{tidyfinance-package}
\alias{tidyfinance}
\alias{tidyfinance-package}
\title{tidyfinance: Helper functions for developers and researchers familiar with Tidy Finance}
\description{
This repository contains an R package that collects helper functions for developers and researchers familiar with [Tidy Financen with R](https://www.tidy-finance.org/r/). The functions provide shortcuts to selected issues that the book disccuses in detail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.tidy-finance.org/r}
  \item \url{https://github.com/tidy-finance/r-tidyfinance}
  \item Report bugs at \url{https://github.com/tidy-finance/r-tidyfinance/issues}
}

}
\author{
\strong{Maintainer}: Christoph Scheuch \email{christoph.scheuch@gmail.com} (\href{https://orcid.org/0009-0004-0423-6819}{ORCID})

}
\keyword{internal}
