% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyfinance-package.R
\docType{package}
\name{tidyfinance-package}
\alias{tidyfinance}
\alias{tidyfinance-package}
\title{tidyfinance: Tidy Finance Helper Functions}
\description{
This package offers helper functions for empirical research in financial economics, addressing a variety of topics covered in Scheuch, Voigt, and Weiss (2023) \doi{10.1201/b23237}. It is designed to provide shortcuts for issues extensively discussed in the book, facilitating easier application of its concepts. For more information and resources related to the book, visit \url{https://www.tidy-finance.org/r/index.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.tidy-finance.org/r/}
  \item \url{https://github.com/tidy-finance/r-tidyfinance}
  \item Report bugs at \url{https://github.com/tidy-finance/r-tidyfinance/issues}
}

}
\author{
\strong{Maintainer}: Christoph Scheuch \email{christoph.scheuch@gmail.com} (\href{https://orcid.org/0009-0004-0423-6819}{ORCID})

Authors:
\itemize{
  \item Stefan Voigt \email{stefan.voigt@econ.ku.dk} (\href{https://orcid.org/0000-0001-5619-3161}{ORCID})
  \item Patrick Weiss \email{patrickw@ru.is} (\href{https://orcid.org/0000-0002-9282-5872}{ORCID})
}

}
\keyword{internal}
